rough implementation of 3eqn melt parametrization
adapted from FESOM with the help of Ralph Timmermann (AWI).
See Holland and Jenkins (1999) for a description.
Variable seawater density needs to be implemented,
it is still a constant.
We also have to clarify if potential or in situ temperatures are the
right input.
The mass_flux field is kidnapped here to get salinity into the code,
this should be fixed.
Maybe a more sophisticated method to calculate the heat capacity of ice
(cpi) can be implemented, it is hardcoded until now.

This is a combination of 4 commits.
raw import of 3eqn melting, basically copy of the stable05 version.

errors corrected. runs.

corrected signs, results look good.

cleanup of meltingparam3eqn.