calculate in situ temperature from pot temperature input.
This is still not effective. in situ temperature
is found through a loop.
Calculation of melting could be restricted to shelves only.

trying to resort and rename, does not help.

trying to resort and rename, does not help.

adlprt working.

adlprt working.

in situ input for oceanmelt3eqn working.

speed could be increased, for example by bounding
calculation only to real shelf area.